Moved all the function defs from InitalCalculations.py over to utils and copied the code from main() to Process Data. Also moved over the functions from ProcessData.py to utils, so the main function to run is ProcessData.py. Consider deleting InitalCalculations.py for clarity's sake
